fix: move process_drep_spo back to governance_state.rs
